2. notes_app_api/app/notes/serializers.py
from .models import Note, Category, Tag
from rest_framework import serializers

class CategorySerializer(serializers.ModelSerializer):
    """Category to front end"""
    class Meta:
        model = Category
        fields = ['id', 'name', 'created_at', 'updated_at']
        read_only_fields = ['id', 'created_at']

class TagSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = ['id', 'name']
        read_only_fields = ['id']

class NotesSerializer(serializers.ModelSerializer):
    category_name = serializers.CharField(
        source='category.name',
        read_only=True,
        default=""
    )
    tags_names = serializers.SlugRelatedField(
        many=True,
        read_only=True,
        slug_field='name',
        source='tags'
    )

    user_name = serializers.CharField(
        source='user.username',
        read_only=True,
    )

    class Meta:
        model = Note
        fields = ['id', 'title', 'content', 'category', 'category_name', 'tags', 'tags_names', 'created_at', 'user', 'user_name']
        read_only_fields = ['id', 'created_at', 'user', 'user_name']

    def create(self, validated_data):
        validated_data['user'] = self.context['request'].user
        return super().create(validated_data)
